fix(web): keep Garmin login routes build-compatible
(cherry picked from commit 93e57f66fd182bb484649a1ee67267722334c9dd)
